uefi-sct/SctPkg: TCG2 Protocol: correct definition of TPMT_HA struct
The TPMT_HA struct defining event log hash algorithms was cut/pasted
from the TCG EFI Protocol specification which used a C struct
with a flexible array member as the last element.  This is incorrect
because TPMT_HA itself is used as an array element, and thus can't
be variable size.

Because the size of hash algorithms varies, this should have been
defined as a union of the sizes of supported hash algorithms.  This is
how is it done in the TPM Library specfication and in EDK2.

In addition, added the SM3 hash type to align with the TPM Library
spec.

Signed-off-by: Stuart Yoder <stuart.yoder@arm.com>